+++ src/Laravel/functions.php
@@ -37,7 +37,9 @@ function rules($rules, array $messages = [], array $customAttributes = []): Clos
         }
 
         $validation = Validator::make(
-            $value !== null ? [$key => $value] : [],
+            $value !== null
+                ? [$key => $value]
+                : [] + $context->data['attributes'] + $context->data['relationships'],
             $validatorRules,
             $messages,
             $customAttributes,